refactor: future-proof the public API with #[non_exhaustive] (0.7.0 hardening)
Mark the public types expected to grow as #[non_exhaustive] so that fixing the
remaining known issues in 0.7.x can add variants/fields without a breaking
change, and retract the orphan TestConfig from the public re-export. Breaking;
the last no-cost window is 0.7.0 (unpublished). Goal: 0.7.0 is the LAST
breaking release for the current known-issue set.

- error.rs: #[non_exhaustive] on RiperfError, ConfigError (future error /
  validation variants, e.g. #45/#100, become additive).
- protocol.rs: #[non_exhaustive] on TransportProtocol (a future transport) and
  on the result model TestResultsJson/StreamResultJson (consumers read these;
  future JSON fields become additive). The -J json_report model was already
  #[non_exhaustive].
- lib.rs: drop TestConfig from the public re-export (server-internal, not used
  in any public signature).
- Migrate test_results_json_structure in-crate (an external test crate can no
  longer construct the now-non_exhaustive result structs).
